refactor: StoreController 클래스 구현
Application 클래스에 존재하는 메인 로직을 분리해내어 구현
